+++ contracts/merkledrop/MerkleDistributor.sol
@@ -5,11 +5,10 @@ pragma solidity 0.8.16;
 import "./MerkleProof.sol";
 
 interface ERC20Like {
-    function transfer(address dst, uint qty) external returns (bool);
+    function transfer(address dst, uint256 qty) external returns (bool);
 }
 
 contract MerkleDistributor {
-
     event Claimed(uint256 index, address account, uint256 amount);
 
     address public immutable token;
@@ -34,19 +33,32 @@ contract MerkleDistributor {
     function _setClaimed(uint256 index) private {
         uint256 claimedWordIndex = index / 256;
         uint256 claimedBitIndex = index % 256;
-        claimedBitMap[claimedWordIndex] = claimedBitMap[claimedWordIndex] | (1 << claimedBitIndex);
+        claimedBitMap[claimedWordIndex] =
+            claimedBitMap[claimedWordIndex] |
+            (1 << claimedBitIndex);
     }
 
-    function claim(uint256 index, address account, uint96 amount, bytes32[] memory merkleProof) external {
-        require(!isClaimed(index), 'MerkleDistributor: Drop already claimed.');
+    function claim(
+        uint256 index,
+        address account,
+        uint96 amount,
+        bytes32[] memory merkleProof
+    ) external {
+        require(!isClaimed(index), "MerkleDistributor: Drop already claimed.");
 
         // Verify the merkle proof.
         bytes32 node = keccak256(abi.encodePacked(index, account, amount));
-        require(MerkleProof.verify(merkleProof, merkleRoot, node), 'MerkleDistributor: Invalid proof.');
+        require(
+            MerkleProof.verify(merkleProof, merkleRoot, node),
+            "MerkleDistributor: Invalid proof."
+        );
 
         // Mark it claimed and send the token.
         _setClaimed(index);
-        require(ERC20Like(token).transfer(account, amount), 'MerkleDistributor: Transfer failed.');
+        require(
+            ERC20Like(token).transfer(account, amount),
+            "MerkleDistributor: Transfer failed."
+        );
 
         emit Claimed(index, account, amount);
     }
