+++ contracts/rescue/UniswapV2Like.sol
@@ -3,27 +3,34 @@
 pragma solidity 0.8.16;
 
 interface UniswapV2RouterLike {
-	function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+    function swapExactTokensForTokens(
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
-	function addLiquidity(
+    function addLiquidity(
         address tokenA,
         address tokenB,
-        uint amountADesired,
-        uint amountBDesired,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountA, uint amountB, uint liquidity);
+        uint256 deadline
+    )
+        external
+        returns (
+            uint256 amountA,
+            uint256 amountB,
+            uint256 liquidity
+        );
 }
 
 interface UniswapV2PairLike {
     function token0() external returns (address);
+
     function token1() external returns (address);
 }
